Moved image utils to separate file
